arrondi correct

+++ src/main/java/pt4/flotsblancs/database/model/CampGround.java
@@ -34,7 +34,7 @@ public class CampGround implements Item {
     @Getter
     @Setter
     @DatabaseField(canBeNull = false, columnName = "price_per_day")
-    private float pricePerDays;
+    private int pricePerDays;
 
     @Getter
     @Setter
